Backend/requirements.txt
fastapi[standard]
pydantic
python-dotenv
motor
langchain-groq
langchain-huggingface
langchain-chroma
langchain-community
langchain-text-splitters
langchain-classic
langchain-core
langchain-mongodb
uvicorn
sentence-transformers
torch
torchvision
pillow
timm
boto3
cloudinary 
python-multipart
huggingface_hub
python-jose[cryptography]
passlib[bcrypt]
